# Implement more methods for `vec_deque::IntoIter`