@Sp00ph 2023-01-03
> Have you benchmarked whether they're better than the default impls? If yes please post before/after results. Vec::IntoIter overrides those methods because it often results in vectorizable code while the default impls don't. E.g. I suspect next_chunk might benefit less than Vec does because it'll see non-const-length copies even on the `Ok` branch while Vec only does that for `Err`.

I don't think there are currently any benchmarks for `vec_deque::IntoIter`. I can try to write a few of those later.