@the8472 2023-01-14
> I can try to write a few of those later.

Yes, please do. If there aren't significant wins we might as well stick with the defaults.

@rustbot author