@the8472 2023-01-17
Hrrm, I don't think this makes sense. fold shouldn't be slower than try_fold. The issue is that fold takes ownership of the iterator and drops it, so you can't reuse the allocation when benchmarking `IntoIter`, you can only do that with borrowing iterators.

In that case we just have to bite the bullet and measure the allocation costs together with the iterator performance.