@Sp00ph 2023-01-17
Right, that makes sense. Makes me think, would it maybe make sense to implement the default `<&mut Iterator>::fold` in terms of `Iterator::try_fold`? That way it couldn't take advantage of all fold optimizations (`try_fold` is still unstable), but it would at least work for most `stdlib` iterators that override `try_fold`.