@Sp00ph 2023-01-17
I changed the `fold` benchmark to use a new deque on every iteration. Now the numbers look like this:
```
master:

test vec_deque::bench_into_iter_fold                     ... bench:       1,979 ns/iter (+/- 160)

This PR:

test vec_deque::bench_into_iter_fold                     ... bench:         395 ns/iter (+/- 16)
```